+++ pl_examples/basic_examples/simple_image_classifier.py
@@ -68,10 +68,10 @@ class LitClassifier(pl.LightningModule):
 
     @staticmethod
     def add_model_specific_args(parent_parser):
-        parser = ArgumentParser(parents=[parent_parser], add_help=False)
+        parser = parent_parser.add_argument_group("LitClassifier")
         parser.add_argument('--hidden_dim', type=int, default=128)
         parser.add_argument('--learning_rate', type=float, default=0.0001)
-        return parser
+        return parent_parser
 
 
 def cli_main():
@@ -105,8 +105,6 @@ def cli_main():
     # ------------
     # testing
     # ------------
-    # todo: without passing model it fails for missing best weights
-    # MisconfigurationException, 'ckpt_path is "best", but ModelCheckpoint is not configured to save the best model.'
     result = trainer.test(model, datamodule=dm)
     pprint(result)